Bug fix: concatenate main.js in last position
The main script shall now run after all constructors are loaded,
not before. This is actually more logical: the main script uses
the constructors and shall thus be loaded after. It used to be
loaded before only because constructors needed the global
namespace this.Coquette defined in the main script.